Change app id from merging to merging2.
