--- app/src/main/java/uk/me/jeremygreen/merging/MainApplication.kt
@@ -1,13 +1,0 @@
-@file:Suppress("unused")
-
-package uk.me.jeremygreen.merging
-
-import android.app.Application
-import com.facebook.drawee.backends.pipeline.Fresco
-
-internal class MainApplication : Application() {
-    override fun onCreate() {
-        super.onCreate()
-        Fresco.initialize(this) // Bitmap caching - https://frescolib.org/
-    }
-}
