--- app/src/main/java/uk/me/jeremygreen/merging/main/screen/MergedImage.kt
@@ -1,27 +1,0 @@
-package uk.me.jeremygreen.merging.main.screen
-
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.res.stringResource
-import uk.me.jeremygreen.merging.R
-
-// id: Long = -3
-// screenName(): String = "MergedImage"
-
-@Composable
-fun MergedImage() {
-    Column (
-        horizontalAlignment = Alignment.CenterHorizontally,
-        verticalArrangement = Arrangement.Center,
-        modifier = Modifier.fillMaxSize()
-    ) {
-        Text(
-            text = stringResource(R.string.merged_image)
-        )
-    }
-}
