--- app/src/main/java/uk/me/jeremygreen/merging/model/Face.kt
@@ -1,14 +1,0 @@
-package uk.me.jeremygreen.merging.model
-
-import androidx.room.Relation
-
-internal data class Face(
-
-    val id: Long,
-
-    val imageId: Long,
-
-    @Relation(parentColumn = "id", entityColumn = "faceId")
-    val coordinates: List<Coordinate>
-
-)
