+++ app/src/main/java/uk/me/jeremygreen/merging2/main/screen/AddImage.kt
@@ -1,0 +1,10 @@
+package uk.me.jeremygreen.merging2.main.screen
+
+import androidx.compose.runtime.Composable
+
+// id: Long = -2
+// screenName(): String = "AddImage"
+@Composable
+fun AddImage() {
+    // Empty - the UI is provided by MainActivity action button.
+}
