+++ app/src/main/java/uk/me/jeremygreen/merging2/model/Coordinate.kt
@@ -1,0 +1,28 @@
+package uk.me.jeremygreen.merging2.model
+
+import androidx.room.Entity
+import androidx.room.ForeignKey
+import androidx.room.Index
+import androidx.room.PrimaryKey
+
+@Entity(
+    tableName = "coordinates",
+    foreignKeys = [
+        ForeignKey(
+            entity=FaceEntity::class,
+            parentColumns = ["id"],
+            childColumns = ["faceId"],
+            onDelete = ForeignKey.CASCADE
+        )
+    ],
+    indices = [Index(value = ["faceId"])]
+)
+internal data class Coordinate(
+    @PrimaryKey(autoGenerate = true)
+    val id: Long,
+
+    val faceId: Long,
+
+    val x: Float,
+    val y: Float
+)
