+++ app/src/main/java/uk/me/jeremygreen/merging2/model/ProcessingStage.kt
@@ -1,0 +1,6 @@
+package uk.me.jeremygreen.merging2.model
+
+object ProcessingStage {
+    const val unprocessed = 1000
+    const val facesDetected = 2000
+}
