+++ fastlane/Appfile
@@ -1,2 +1,2 @@
 json_key_file("google-play-service.json")
-package_name("uk.me.jeremygreen.merging")
+package_name("uk.me.jeremygreen.merging2")
